backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
llama-cpp-python
